requirements.txt
Flask==2.1.2
GitPython==3.1.31
werkzeug==2.0.3